Crate validate_npm_package_name

A Rust implementation of the validation rules from the core JS package validate-npm-package-name.

Enums

Validity

Constants

BLACKLIST 🔒

BUILTINS 🔒

Statics

ENCODE_URI_SET 🔒

The set of characters to encode, matching the characters encoded by encodeURIComponent

SCOPED_PACKAGE 🔒

SPECIAL_CHARS 🔒

Functions

done 🔒

validate