Crate validate_npm_package_name

A Rust implementation of the validation rules from the core JS package validate-npm-package-name.

Structs

BLACKLIST

BUILTINS

SCOPED_PACKAGE

SPECIAL_CHARS

Enums

Validity

Statics

ENCODE_URI_SET

The set of characters to encode, matching the characters encoded by encodeURIComponent

Functions

done

validate